+++ ChangeLog.md
@@ -4,6 +4,7 @@
 - Updated using `verida-js/vda-common` package
 - Removed unnecessary codes
 - Updated test codes
+- Updated contract interaction. Removed usage of `ContractFactory` and used `Contract` in `ethers`
 
 V(1.0.0)
 --------------------